+++ mod-source-record-manager-server/src/main/java/org/folio/services/ChangeEngineServiceImpl.java
@@ -146,7 +146,7 @@ public class ChangeEngineServiceImpl implements ChangeEngineService {
       .onSuccess(parsedRecords -> {
         fillParsedRecordsWithAdditionalFields(parsedRecords);
 
-        if (updateMarcActionExists(jobExecution)) {
+        if (updateMarcActionExists(jobExecution) || updateInstanceActionExists(jobExecution)) {
           updateRecords(parsedRecords, jobExecution, params)
             .onSuccess(ar -> promise.complete(parsedRecords))
             .onFailure(promise::fail);
@@ -188,7 +188,7 @@ public class ChangeEngineServiceImpl implements ChangeEngineService {
   }
 
   private Future<Boolean> updateRecords(List<Record> records, JobExecution jobExecution, OkapiConnectionParams params) {
-    LOGGER.info("Records have not been saved in record-storage, because job contains action for Marc update");
+    LOGGER.info("Records have not been saved in record-storage, because job contains action for Marc or Instance update");
     return recordsPublishingService
       .sendEventsWithRecords(records, jobExecution.getId(), params, DI_MARC_FOR_UPDATE_RECEIVED.value());
   }
@@ -229,6 +229,13 @@ public class ChangeEngineServiceImpl implements ChangeEngineService {
       Action.UPDATE);
   }
 
+  private boolean updateInstanceActionExists(JobExecution jobExecution) {
+    return containsMarcActionProfile(
+      jobExecution.getJobProfileSnapshotWrapper(),
+      List.of(FolioRecord.INSTANCE),
+      Action.UPDATE);
+  }
+
   private boolean deleteMarcActionExists(JobExecution jobExecution) {
     return containsMarcActionProfile(
       jobExecution.getJobProfileSnapshotWrapper(),
@@ -237,13 +244,13 @@ public class ChangeEngineServiceImpl implements ChangeEngineService {
   }
 
   private boolean containsMarcActionProfile(ProfileSnapshotWrapper profileSnapshot,
-                                                  List<FolioRecord> records, Action action) {
+                                                  List<FolioRecord> entityTypes, Action action) {
     List<ProfileSnapshotWrapper> childWrappers = profileSnapshot.getChildSnapshotWrappers();
     for (ProfileSnapshotWrapper childWrapper : childWrappers) {
       if (childWrapper.getContentType() == ProfileSnapshotWrapper.ContentType.ACTION_PROFILE
-        && actionProfileMatches(childWrapper, records, action)) {
+        && actionProfileMatches(childWrapper, entityTypes, action)) {
         return true;
-      } else if (containsMarcActionProfile(childWrapper, records, action)) {
+      } else if (containsMarcActionProfile(childWrapper, entityTypes, action)) {
         return true;
       }
     }